requirements.txt
alembic==1.0.0
atomicwrites==1.1.5
attrs==18.1.0
click==6.7
Flask==0.12.2
Flask-Migrate==2.0.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.2.0
nose==1.3.7
pluggy==0.6.0
py==1.5.4
PyJWT==1.6.4
pytest==3.6.3
python-dateutil==2.7.3
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.1.10
SQLAlchemy-Utils==0.32.14
Werkzeug==0.14.1
